@@ -1,9 +1,8 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import { createStore } from 'vuex';
 import { auth } from './auth.module';
-Vue.use(Vuex);
-export default new Vuex.Store({
+
+export default createStore({
   modules: {
-    auth
+    auth,
   }
 });
